refactor: 탈퇴 제외 controller-mapper-service 분리 및 메소드명 리네이밍 #236
