+++ src/main/java/com/ceos/bankids/service/UserService.java
@@ -1,29 +1,23 @@
 package com.ceos.bankids.service;
 
-import com.ceos.bankids.mapper.request.ExpoRequest;
-import com.ceos.bankids.mapper.request.UserTypeRequest;
 import com.ceos.bankids.domain.User;
-import com.ceos.bankids.dto.LoginDTO;
-import com.ceos.bankids.dto.MyPageDTO;
 import com.ceos.bankids.dto.OptInDTO;
 import com.ceos.bankids.dto.UserDTO;
+import com.ceos.bankids.mapper.request.ExpoRequest;
+import com.ceos.bankids.mapper.request.UserTypeRequest;
 import java.util.Optional;
 import org.springframework.stereotype.Service;
 
 @Service
 public interface UserService {
 
-    public Optional<User> getUserByAuthenticationCodeNullable(String code);
+    public Optional<User> readUserByAuthenticationCodeNullable(String code);
 
-    public User postNewUser(String username, String code, String provider);
+    public User createNewUser(String username, String code, String provider);
 
     public UserDTO updateUserType(User user, UserTypeRequest userTypeRequest);
 
-    public LoginDTO issueNewTokens(User user, String newAccessToken, String newRefreshToken);
-
-//    public void setNewCookie(User user, HttpServletResponse response);
-
-    public MyPageDTO getUserInformation(User user);
+    public User updateRefreshToken(User user, String newRefreshToken);
 
     public void updateUserLogout(User user);
 
@@ -35,5 +29,5 @@ public interface UserService {
 
     public OptInDTO updateServiceOptIn(User user);
 
-    public OptInDTO getOptIn(User user);
+    public OptInDTO readOptIn(User user);
 }
